Process the iterative registration
- Got the transformation matrix
- Need to initialize the array which is used to save transformed image
- Need to change that it don't use max(0,value) , cause it makes the image too bright

+++ IP_project/IP_project/ip_registration.h
@@ -30,7 +30,7 @@ public:
 private:
 	void calculateCOI(mc::image3d<TYPE>* img, Point* coi);
 	double calculateSimilarity(std::vector<Point>& moved);
-	Eigen::Matrix4d makeTransformMatrix(int t_param,int degree);
+	Eigen::Matrix4d makeTransformMatrix(int t_param,double degree);
 	void transform(std::vector<Point>* target, Eigen::Matrix4d mat);
 
 	mc::image3d<TYPE>* m_ref;
